requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.4
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
Pillow==8.0.1
pytz==2020.4
soupsieve==2.1
sqlparse==0.4.1
